Transfer: Double to BigDecimal refactoring for amount

+++ transfer/src/main/java/ru/yandex/practicum/transfer/dto/AccountsRequest.java
@@ -1,4 +1,6 @@
 package ru.yandex.practicum.transfer.dto;
 
-public record AccountsRequest(Long userId, String currency, Double amount) {
+import java.math.BigDecimal;
+
+public record AccountsRequest(Long userId, String currency, BigDecimal amount) {
 }
