+++ transfer/src/main/java/ru/yandex/practicum/transfer/dto/TransferRequest.java
@@ -1,8 +1,10 @@
 package ru.yandex.practicum.transfer.dto;
 
+import java.math.BigDecimal;
+
 public record TransferRequest(Long fromUserId,
                               Long toUserId,
                               String fromCurrency,
                               String toCurrency,
-                              Double amount) {
+                              BigDecimal amount) {
 }
