+++ transfer/src/main/java/ru/yandex/practicum/transfer/service/NotificationsClient.java
@@ -7,6 +7,8 @@ import org.springframework.web.client.RestTemplate;
 import ru.yandex.practicum.transfer.dto.NotificationRequest;
 import ru.yandex.practicum.transfer.dto.NotificationResponse;
 
+import java.math.BigDecimal;
+
 @Service
 @RequiredArgsConstructor
 @Slf4j
@@ -15,7 +17,7 @@ public class NotificationsClient {
     private final RestTemplate restTemplate;
 
     public void sendNotification(Long userId, String type, String message,
-                                 Double amount, String currency) {
+                                 BigDecimal amount, String currency) {
         log.debug("Sending notification to user {}: {}", userId, type);
 
         NotificationRequest request = new NotificationRequest(userId, type, message, amount, currency);
